+++ src/Subsystems/DriveBaseSubsystem.h
@@ -15,7 +15,8 @@ private:
 public:
 	DriveBaseSubsystem();
 	void InitDefaultCommand() override;
-	void MecanumDrive(float x, float y, float rot, float gyro);
+	void MecanumDrive(double x, double y, double rot, double gyro);
+	//void MecanumDrive(frc::XboxController*joystick);
 };
 
 #endif  // DriveBaseSubsystem_H
